Added a lifecycle widget
